Use value attribute instead of event.data
In the input event listeners for 'pass' and
'checkPass', the variables 'passInputs' and
'checkPassInputs' will now be set to their
respective input field's value attribute instead
of the input event's data property. This will
ensure the fields are still validated even if the
user deletes data, as the event.data property
returns null if the user hits backspace.